[msbuild] Merge the FindWatchOS2AppBundle and FindWatchOS2AppBundleTaskBase classes.
We no longer need two have overridable logic for remote builds, so the
non-abstract task class and the abstract base class can be merged.

Also enable nullability and fix any issues.
