--- msbuild/Xamarin.iOS.Tasks/Tasks/FindWatchOS2AppBundle.cs
@@ -1,24 +1,0 @@
-using Microsoft.Build.Framework;
-using Xamarin.Messaging.Build.Client;
-
-namespace Xamarin.iOS.Tasks {
-	public class FindWatchOS2AppBundle : FindWatchOS2AppBundleTaskBase, ICancelableTask {
-		public override bool Execute ()
-		{
-			if (!ShouldExecuteRemotely ())
-				return base.Execute ();
-
-			var taskRunner = new TaskRunner (SessionId, BuildEngine4);
-
-			taskRunner.FixReferencedItems (WatchAppReferences);
-
-			return taskRunner.RunAsync (this).Result;
-		}
-
-		public void Cancel ()
-		{
-			if (ShouldExecuteRemotely ())
-				BuildConnection.CancelAsync (BuildEngine4).Wait ();
-		}
-	}
-}
