+++ msbuild/Xamarin.iOS.Tasks/Tasks/FindWatchOS2AppBundleTaskBase.cs
@@ -5,26 +5,35 @@ using Microsoft.Build.Utilities;
 
 using Xamarin.MacDev.Tasks;
 using Xamarin.MacDev;
+using Xamarin.Messaging.Build.Client;
 using Xamarin.Utils;
 
+#nullable enable
+
 namespace Xamarin.iOS.Tasks {
-	public abstract class FindWatchOS2AppBundleTaskBase : XamarinTask {
+	public class FindWatchOS2AppBundle : XamarinTask, ICancelableTask {
 		#region Inputs
 
 		[Required]
-		public ITaskItem [] WatchAppReferences { get; set; }
+		public ITaskItem [] WatchAppReferences { get; set; } = Array.Empty<ITaskItem> ();
 
 		#endregion Inputs
 
 		#region Outputs
 
 		[Output]
-		public string WatchOS2AppBundle { get; set; }
+		public string WatchOS2AppBundle { get; set; } = string.Empty;
 
 		#endregion
 
 		public override bool Execute ()
 		{
+			if (ShouldExecuteRemotely ()) {
+				var taskRunner = new TaskRunner (SessionId, BuildEngine4);
+				taskRunner.FixReferencedItems (WatchAppReferences);
+				return taskRunner.RunAsync (this).Result;
+			}
+
 			var pwd = PathUtils.ResolveSymbolicLinks (Environment.CurrentDirectory);
 
 			if (WatchAppReferences.Length > 0) {
@@ -34,6 +43,12 @@ namespace Xamarin.iOS.Tasks {
 			}
 
 			return !Log.HasLoggedErrors;
+		}
+
+		public void Cancel ()
+		{
+			if (ShouldExecuteRemotely ())
+				BuildConnection.CancelAsync (BuildEngine4).Wait ();
 		}
 	}
 }
